requirements.txt
google-adk
pytest
pytest-asyncio
deprecated
pandas
tabulate
tqdm
scikit-learn
mcp
google-adk-mcp
python-jobspy